requirements.txt
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2
pathlib>=1.0.1
